+++ quantum_unity/Assets/Scripts/Fusion Fighters/Controllers/WeaponController.cs
@@ -168,7 +168,7 @@ public class WeaponController : Controller<WeaponController>
 
         if (_price.isActiveAndEnabled)
         {
-            ulong price = template.Price;
+            uint price = template.Price;
             _price.SetText($"${price}");
 
             _price.color = InventoryController.Instance.HasEnoughCurrency(price) ? Color.white : Color.red;
@@ -185,7 +185,7 @@ public class WeaponController : Controller<WeaponController>
 
         if (_price.isActiveAndEnabled)
         {
-            ulong price = _template.Price + material.Price;
+            uint price = _template.Price + material.Price;
             _price.SetText($"${price}");
 
             _price.color = InventoryController.Instance.HasEnoughCurrency(price) ? Color.white : Color.red;
@@ -202,7 +202,7 @@ public class WeaponController : Controller<WeaponController>
 
         if (_price.isActiveAndEnabled)
         {
-            ulong price = _template.Price + _material.Price + enhancer.Price;
+            uint price = _template.Price + _material.Price + enhancer.Price;
             _price.SetText($"${price}");
 
             _price.color = InventoryController.Instance.HasEnoughCurrency(price) ? Color.white : Color.red;
