+++ quantum_unity/Assets/Scripts/Fusion Fighters/Save Data/Inventory.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public struct Inventory
 {
-    public ulong Currency;
+    public uint Currency;
 
     [Tooltip("Dictionary of all Info Assets to a tuple which represents if they start off unlocked and how many you have (null means infinite)")] public Dictionary<AssetRefInfoAsset, Tuple<bool, Nullable<int>>> ItemCollection;
